+++ economy/admin.py
@@ -9,6 +9,7 @@ from economy.models import (
     Transfer,
     SociOrderSession,
     SociOrderSessionOrder,
+    ExternalCharge,
 )
 
 
@@ -16,6 +17,11 @@ from economy.models import (
 class SociBankAccountAdmin(admin.ModelAdmin):
     list_display = ["user", "card_uuid", "balance"]
     readonly_fields = ["balance"]
+    search_fields = [
+        "user__username",
+        "user__first_name",
+        "user__last_name",
+    ]
 
 
 @admin.register(SociProduct)
@@ -73,4 +79,9 @@ class SociOrderSessionAdmin(admin.ModelAdmin):
 
 @admin.register(SociOrderSessionOrder)
 class UserSociOrderSessionCollectionAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(ExternalCharge)
+class ExternalChargeAdmin(admin.ModelAdmin):
     pass